Refactor(folder): 폴더 정리
