+++ src/main/java/com/example/backend/controller/BANK/AccountController.java
@@ -1,7 +1,7 @@
-package com.example.backend.controller;
+package com.example.backend.controller.BANK;
 
 import com.example.backend.dto.account.*;
-import com.example.backend.service.AccountService;
+import com.example.backend.service.BANK.AccountService;
 import com.example.backend.service.CardService;
 
 import lombok.RequiredArgsConstructor;
@@ -12,8 +12,6 @@ import org.springframework.web.bind.annotation.*;
 
 import java.math.BigDecimal;
 import java.time.YearMonth;
-import java.util.List;
-import java.util.Map;
 
 @RestController
 @RequestMapping("/api/account")
@@ -26,12 +24,12 @@ public class AccountController {
 
     // accountNum을 통해 account 정보 가져오기
     // bank 에서 사용
-    @GetMapping("/info")
-    public ResponseEntity<AccountInfoDTO> getInfo(
-            @RequestParam String accountNum) {
-        AccountInfoDTO accountInfo = accountService.getInfo(accountNum);
-        return ResponseEntity.ok(accountInfo);
-    }
+//    @GetMapping("/info")
+//    public ResponseEntity<AccountInfoDTO> getInfo(
+//            @RequestParam String accountNum) {
+//        AccountInfoDTO accountInfo = accountService.getInfo(accountNum);
+//        return ResponseEntity.ok(accountInfo);
+//    }
 
     // 지출 간단 보기
     @GetMapping("/expense")
