+++ src/main/java/com/example/backend/repository/GoalsRepository.java
@@ -1,6 +1,6 @@
 package com.example.backend.repository;
 
-import com.example.backend.model.Goals;
+import com.example.backend.model.BUSINESS.Goals;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface GoalsRepository extends JpaRepository<Goals, Long> {
